requirements/dev.txt
coverage==4.5.3
Django==2.1.3
django-appconf==1.0.3
django-ckeditor==5.7.1
django-imagekit==4.0.2
django-js-asset==1.2.2
entrypoints==0.3
factory-boy==2.12.0
Faker==2.0.0
flake8==3.7.8
mccabe==0.6.1
pilkit==2.0
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.0
pytz==2018.7
six==1.12.0
text-unidecode==1.2